rename controller namespace to humbleObject
